fix: location deletion is user-specific

+++ src/apps/weather/services/location_service.py
@@ -21,4 +21,4 @@ class LocationService:
 
     @staticmethod
     def delete_location(dto: LocationDTO) -> QuerySet:
-        return Location.objects.filter(latitude=dto.latitude, longitude=dto.longitude).delete()
+        return Location.objects.filter(user_id=dto.user).get(latitude=dto.latitude, longitude=dto.longitude).delete()
